Restructured to be more efficient 🚀
