+++ src/main/cpp/Basic/Component.cpp
@@ -1,19 +1,5 @@
 #include <Basic/Component.h>
 
-Component::~Component() = default;
-
-void Component::doPersistantConfiguration() { }
-
-void Component::resetToMatchMode(MatchMode mode) { }
-
-void Component::sendFeedback() { }
-
-void Component::sendFeedback() { }
-
-void Component::teleOpProcess() { }
-
-void Component::autoProcess() { }
-
 Component::MatchMode Component::getMatchMode() {
     if(frc::DriverStation::IsDisabled()) {
         return MatchMode::DISABLED;
@@ -30,10 +16,6 @@ Component::MatchMode Component::getMatchMode() {
 
 Component::MatchMode Component::getLastMode() {
     return lastMode;
-}
-
-units::ampere_t Component::getCurrent() {
-    return 0_A;
 }
 
 void Component::callResetToMode(MatchMode _lastMode) {
